renaming modules to avoid including hanabi_ai in title
PEP recommends short names for modules. with absolute imports,
modules are unambiguous even without hanabi_ai in their names
eliminating redundant naming and updating imports

Imports look prettier now

Finally, this makes play_game less of an exceptional name
