+++ hanabi_ai/model/game_info.py
