+++ hanabi_ai/model/hand.py
