+++ hanabi_ai/tests/game_info_tests.py
@@ -1,5 +1,5 @@
 import unittest
-from hanabi_ai.model.hanabi_game_info import GameInfo
+from hanabi_ai.model.game_info import GameInfo
 
 class GameInfoTests(unittest.TestCase):
     def setUp(self):
