+++ hanabi_ai/tests/hand_tests.py
@@ -1,6 +1,6 @@
 import unittest
-from hanabi_ai.model.hanabi_hand import HanabiHand
-from hanabi_ai.model.hanabi_card import HanabiCard, HanabiColor
+from hanabi_ai.model.hand import HanabiHand
+from hanabi_ai.model.card import HanabiCard, HanabiColor
 
 class HanabiHandTests(unittest.TestCase):
     def setUp(self):
