+++ hanabi_ai/tests/table_variant_1_tests.py
@@ -1,7 +1,7 @@
 import unittest
-from hanabi_ai.model.hanabi_table import HanabiTable
-from hanabi_ai.model.hanabi_card import HanabiCard, HanabiColor
-from hanabi_ai.model.hanabi_deck import HanabiVariant
+from hanabi_ai.model.table import HanabiTable
+from hanabi_ai.model.card import HanabiCard, HanabiColor
+from hanabi_ai.model.deck import HanabiVariant
 
 class HanabiTableVariant1Tests(unittest.TestCase):
     def setUp(self):
